+++ eggs/naaya.content.meeting/naaya/content/meeting/participants.py
@@ -10,8 +10,7 @@ from DateTime import DateTime
 
 # Naaya imports
 from Products.NaayaCore.FormsTool.NaayaTemplate import NaayaPageTemplateFile
-from Products.NaayaCore.managers.import_export import (generate_excel,
-                                                       generate_csv)
+from Products.NaayaCore.managers.import_export import generate_excel
 
 # Meeting imports
 from naaya.content.meeting import (WAITING_ROLE, PARTICIPANT_ROLE,
@@ -489,39 +488,6 @@ class Participants(SimpleItem):
         RESPONSE.setHeader('Content-Disposition', 'attachment; filename=%s'
                            % filename)
         return generate_excel(header, rows)
-
-    security.declareProtected(view, 'download_webex')
-
-    def download_webex(self, REQUEST=None, RESPONSE=None):
-        """exports the participants listing in a Webex-compatible CSV file"""
-
-        # This will be refactored to export to Skype, Webex will be shut
-        # down. Right now this method is not used.
-
-        header = ['UUID', 'Name', 'Email', 'Company', 'JobTitle', 'URL',
-                  'OffCntry', 'OffLocal', 'CellCntry', 'CellLocal', 'FaxCntry',
-                  'FaxLocal', 'Address1', 'Address2', 'City', 'State/Province',
-                  'Zip/Postal', 'Country', 'Time Zone', 'Language', 'Locale',
-                  'UserName', 'Notes']
-        rows = []
-        participants = self.getAttendees()
-        for participant in participants:
-            part_info = self.getAttendeeInfo(participant)
-            participant_info = [
-                part_info['uid'], part_info['name'], part_info['email'],
-                part_info['organization'], part_info['role'],
-                '', part_info['phone'], '', '', '', '', '', '', '', '', '', '',
-                get_country_name(part_info['country']),
-                '', '', '', part_info['uid'], '']
-            rows.append(participant_info)
-
-        filename = '%s_%s_%s.csv' % (self.getMeeting().getId(), self.id,
-                                     datetime.now().strftime(
-                                     "%Y-%m-%d_%H-%M-%S"))
-        RESPONSE.setHeader('Content-Type', 'text/csv')
-        RESPONSE.setHeader('Content-Disposition', 'attachment; filename=%s'
-                           % filename)
-        return generate_csv(header, rows)
 
 
 InitializeClass(Participants)